feat: :Mix user command
This merges in the work by @brendalf in https://github.com/brendalf/mix.nvim.

Closes #15
