+++ lua/elixir/mix/exs.lua
@@ -1,0 +1,18 @@
+local git = require("elixir.mix.git")
+
+local M = {}
+
+function M.path_mix_exs()
+	local git_workdir_path = git.workdir_path()
+	local git_mix_exs_path = git.find_file("mix.exs")
+	local mix_exs_fullpath = table.concat({ git_workdir_path, git_mix_exs_path }, "/")
+
+	local file_exists = not vim.tbl_isempty(vim.loop.fs_stat(mix_exs_fullpath) or {})
+	return {
+		mix_file = mix_exs_fullpath,
+		mix_dir = vim.fn.fnamemodify(mix_exs_fullpath, ":p:h"),
+		file_exists = file_exists,
+	}
+end
+
+return M
